Clean NOAA earthquake data

This function takes raw NOAA earthquake data and returns a clean data frame. The clean data frame includes a date column created by uniting the year, month, and day, converting it to the Date class. It also converts the LATITUDE and LONGITUDE columns to numeric class.

eq_clean_data(path_of_data = "inst/extdata/earthquakes.tsv")

Arguments

path_of_data

character string containing path of .tsv file for NOAA earthquake data.

Value

A clean data frame with date, latitude, and longitude columns.

Examples

if (FALSE) {
data <- read.csv(system.file("extdata", "your_data.csv", package = "visualizeNOAA"))
eq_clean_data(data)
}